+++ tests/console/conftest.py
@@ -1,5 +1,4 @@
 import os
-import shutil
 
 import pytest
 
@@ -13,61 +12,15 @@ from poetry.poetry import Poetry as BasePoetry
 from poetry.repositories import Pool
 from poetry.repositories import Repository as BaseRepository
 from poetry.repositories.exceptions import PackageNotFound
-from poetry.utils._compat import PY2
-from poetry.utils._compat import WINDOWS
 from poetry.utils._compat import Path
 from poetry.utils.toml_file import TomlFile
-from poetry.vcs.git import ParsedUrl
-
-
-try:
-    import urllib.parse as urlparse
-except ImportError:
-    import urlparse
+from tests.helpers import mock_clone
+from tests.helpers import mock_download
 
 
 @pytest.fixture()
 def installer():
     return NoopInstaller()
-
-
-def mock_clone(self, source, dest):
-    # Checking source to determine which folder we need to copy
-    parsed = ParsedUrl.parse(source)
-
-    folder = (
-        Path(__file__).parent.parent
-        / "fixtures"
-        / "git"
-        / parsed.resource
-        / parsed.pathname.lstrip("/").rstrip(".git")
-    )
-
-    shutil.rmtree(str(dest))
-    shutil.copytree(str(folder), str(dest))
-
-
-def mock_download(self, url, dest):
-    parts = urlparse.urlparse(url)
-
-    fixtures = Path(__file__).parent.parent / "fixtures"
-    fixture = fixtures / parts.path.lstrip("/")
-
-    if dest.exists():
-        shutil.rmtree(str(dest))
-
-    # Python2 does not support os.symlink on Windows whereas Python3 does.  os.symlink requires either administrative
-    # privileges or developer mode on Win10, throwing an OSError is neither is active.
-    if WINDOWS:
-        if PY2:
-            shutil.copyfile(str(fixture), str(dest))
-        else:
-            try:
-                os.symlink(str(fixture), str(dest))
-            except OSError:
-                shutil.copyfile(str(fixture), str(dest))
-    else:
-        os.symlink(str(fixture), str(dest))
 
 
 @pytest.fixture
