+++ tests/fixtures/with_local_config/poetry.toml
@@ -1,2 +1,3 @@
 [virtualenvs]
+in-project = false
 create = false
